(fix) change the sigio handler of perf_test
It now supports output K samples, K is the sampling period given by the programmer.

Need to fix: there is a bug when read addresses out and build the memory region. This bug
is caused by reading the NULL void * value out and cast it to the uint64_t * then dereference
this NULL pointer.